requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==5.0.0
atomicwrites==1.4.1
attrs==24.2.0
billiard==4.2.1
black==24.10.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==3.2.16
django-filter==24.3
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.8
flake8==7.1.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
kombu==5.4.2
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
Pillow==9.3.0
platformdirs==4.3.6
pluggy==0.13.1
prompt_toolkit==3.0.48
py==1.11.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
requests==2.26.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
toml==0.10.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
